+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 ## [Pending Release](https://github.com/lightstep/otel-launcher-java/compare/0.14.0...master)
 * Use OpenTelemetry Java API version 0.16.0
 * Use OpenTelemetry Java Instrumentation version 0.16.0
-* Remove otel.exporter.otlp.span.insecure system property and OTEL_EXPORTER_OTLP_SPAN_INSECURE environment variable
+* Deprecate otel.exporter.otlp.span.insecure system property and OTEL_EXPORTER_OTLP_SPAN_INSECURE environment variable
 * Require span endpoint to start with http:// or https://
 
 <a name="0.14.0"></a>